Major architecture redesign - version 0.3.0
This changes the architecture of `configure_me` to use two crates - one
for codegen, called `configure_me_codegen` from now and a new
`configure_me`, which serves as a prelude/facade.

This combined with other changes (e.g. `include_config!()` macro) allows
us to shrink boilerplate greatly. The crate using it is much easier to
setup and maintain.

Further, this crate replaces `FromStr` with `ParseArg`, which allows
using `PathBuf` and `OsString` types, thus allowing to work with
weirdly-encoded paths. In the future, it'll be possible to implement
even nicer error messages. Finally, conversion to `String` avoids
allocation.