Refactored `merge_in` to swap its arguments
The order of arguments of `merge_in` was unintuitive and impractical
because `&mut self` should refer to existing aggregating configuration
while `other` should refer to new configuration. It was backwards.

This resolves the issue by swapping the arguments but makes sure to not
break outside interface, by artificailly swapping when there are
multiple config files. A breaking change is expected in the future.

Fixes #31