Man page generation exported as a separate fn
Generating man pages was previously tied to running build scripts. This
prevented external tools to use them separately, e.g. for packaging.

In order to improve the situation, build-helper had to be changed for
`cargo_toml`, because it was panicking when env vars were missing.
`cargo_toml` additionally allow us to define the path to the specification
in package.metadata of Cargo.toml. This will make it possible for other
tools to find it and avoid hard-coding into build scripts.