Added documentation about derive
The question of derives comes up fairly often and for a good reason.
Unfortunately the response is "there are good reasons why it's
difficult". Since this wasn't documented, it might have led to people
dismissing the crate as weird without bothering to ask.

So this change documents the issue for newcomers to see immediately.