Change sequence to map
Writing `name = "foo"` was a bit verbose without significant value added
so this changes it to `[param.foo]` od `[switch.foo]` by changing
sequence into map.

It still internally uses a `Vec` to preserve the order making the
compilation deterministic.

A notable downside of this change is degradation in error messages if
one duplicates a key among same map - two same-named `param`s or
`switch`es. I believe it's still worth it since duplicate keys are easy
to identify anyway and maybe it'll be possible to improve this in the
future.

Closes #47