Fix and improve `require_program_name` option
Originally, `program_name` was meant to be configurable in previous
commit. This was forgotten so the option had no effect. Additionally
it's more useful to also have the option of not adding it into the
`Metadata` struct. Especially once visibility changes to `pub(crate)`
because we don't want to warn about that.

This commit changes the option to be named just `program_name` and take
one of three strings: `unused` (default), `optional`, `required`. Then
it modifies codegen to make the field and its type conditional. The
appropriate tests are added as well.