Remove useless conversions
Clippy was complaining that the generated code contained useless
conversions. This was originally intentional to make the code simpler.
The lint could just be `#[allow]`ed but actually removing useless
conversions has the advantage that the user who accidentally used
`convert_into` on the same type would get the warning.

Closes #49