Don't require `#[macro_use]` on `configure_me`
Previously there was a requirement to add
`#[macro_use] extern crate configure_me` in the crate which made usage
annoying but supported older versions of Rust which were unable to use
paths in macro invocations. Such old versions of Rust are no longer
supported therefore we can safely drop the requirement to have that
declaration.